fix(api): update test script to run all tests and refine limit validation in game account router; enhance elo tracking logic with unique constraint for match history

+++ packages/api/src/routers/gameAccount.ts
@@ -93,7 +93,7 @@ export const gameAccountRouter = router({
   getCs2FaceitMatchHistory: protectedProcedure
     .input(
       gameAccountIdSchema.extend({
-        limit: z.number().min(1).max(100).default(40),
+        limit: z.number().int().min(1).max(100).default(40),
       }),
     )
     .use(requireGameAccountAccess("public-read"))
